+++ elmyra.ip.access.epo/CHANGES.rst
@@ -7,6 +7,7 @@ development
 - middleware: reverse kindcode fixing for DE documents at DEPATISconnect
 - middleware: enhance patent number normalization for old US publications, e.g. US000000024087E => USRE24087E
 - middleware: perform kindcode fixing also on patent normalization api
+- ui: move “Fetch result numbers” button to results tool menu
 
 0.86.2
 ======